BnR rook-ceph: fix for IPv6 network
When rook-ceph recovery is performed on a system that uses
an IPv6 network, the clusterIP is not parsed correctly,
causing communication failures and affecting the cluster's
recovery.

To resolve this, the IP family (IPv4 or IPv6) of the
clusterIP is checked, and when it is IPv6, the necessary
adjustments are made in relation to the brackets.

Additionally, a final ceph status check has also been added.
Because depending on the system configuration and the status
of the pods during recovery, MDS failures may occur, which
will prevent ceph from becoming HEALTH_OK. Therefore, it is
checked whether the ceph status is 'mds daemon damaged' or
'filesystem is degraded'. If so, the cephfs recovery process
is run again to fix this.

A task has also been added to the playbook to check
health of Kubernetes before starting rook-ceph recovery.

Test Plan:
- PASS: optimized SX B&R with IPv4
- PASS: legacy DX B&R with IPv4
- PASS: legacy DX B&R with IPv6

Closes-Bug: 2098747

Change-Id: Icea44a66e01e2994d14baca135f9e0115f5f6f7e
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>