Fix: Update optimized restore to use hostname
The recent change in hieradata filename from <ipaddress>.yaml
to <hostname>.yaml caused LV sizes to not be correctly
restored. This commit modifies the logic to handle the new
filename format and ensures that the restore process fails
if the file does not exist.

Test Plan
PASS: Optimized restore on AIO-SX, stx10
PASS: Optimized restore on AIO-SX subcloud, stx10

Closes-Bug: 2101882

Change-Id: I9ac778e3b01e104ef6bb5e468716563141f78ff1
Signed-off-by: Sai Lakshmi Teja Vanka <SaiLakshmiTeja.Vanka@windriver.com>