Check cephfs recovery commands during rook-ceph restore
If the open deployment model is used and depending on how the
monitors and OSDs were distributed in the cluster, the 'cephfs'
commands may hang during cephfs recovery due to missing data
since only controller-0 is running at that time.

To solve this, the remaining cephfs recovery commands are not
executed if the first command fails.

This will not affect the rook-ceph restore, as there is a cephfs
check at the end of the process, when all hosts are online,
and the recovery is redone.

Test Plan:
- PASS: B&R on STD w/ 3 workers
        open deployment model and 1 mon and 1 OSD on each host.

Closes-Bug: 2106479

Change-Id: Ifa136dfe76bd9ee76e346f86f097f66dbfdff463
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>