Configure gateway when it is defined for mgmt and admin
Eliminate the use of role identification to define the gateways for the
mgmt and admin networks. Gateways are now configured whenever they are
explicitly defined, regardless of the node's role. Additionally,
configure the system controller network when it is defined. There is no
need to check the role or determine whether the node is a subcloud in
this case either. The presence of the system controller network itself
is sufficient to proceed with the configuration.

Test Plan
=========

[PASS] Bootstrap DC dual-stack system with admin network and gateway
defined. Verify that admin IPv4 and IPv6 gateways are set in the
addrpool table and in the database's addresses table as
controller-gateway-admin.

[PASS] Bootstrap DC dual-stack system with mgmt network and gateway, but
no admin network. Verify that mgmt IPv4 and IPv6 gateways are set in the
addrpool table and in the database's addresses table as
controller-gateway-mgmt.

[PASS] Bootstrap DC single-stack system with mgmt network and gateway,
but no admin network. Verify that mgmt IPv4 gateway is set in the
addrpool table and in the database's addresses table as
controller-gateway-mgmt.

[PASS] Bootstrap SX system without defining mgmt gateway during
bootstrap. Verify that no mgmt gateway is set in the addrpool table or
in the database's addresses table.

[PASS] Bootstrap SX dual-stack system with mgmt gateway defined during
bootstrap. Verify that mgmt IPv4 and IPv6 gateways are set in the
addrpool table and in the database's addresses table.

[PASS] Bootstrap SX dual-stack system with admin network and gateway
defined. Verify that admin IPv4 and IPv6 gateways are set in the
addrpool table and in the database's addresses table.

[PASS] Bootstrap SX dual-stack system with admin network and gateways
defined for both admin and mgmt. Verify that both admin and mgmt
gateways are set in the addrpool table and in the database's addresses
table.

[PASS] Bootstrap SX dual-stack system with admin network and gateway
defined. Verify that admin IPv4 and IPv6 gateways are set in the
addrpool table and in the database's addresses table.

[PASS] Bootstrap SX dual-stack system with admin network (no gateway)
and mgmt network with gateway defined. Verify that admin IPv4 and IPv6
gateways are not set in the addrpool or database addresses table.
Verify that mgmt IPv4 and IPv6 gateways are correctly set in both.

Story: 2011312
Task: 51990

Change-Id: I6d5ff92ace8b52a61ac44ff4660936e551c07f35
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>