Add support for Kubernetes 1.32.2
This adds support for Kubernetes 1.32.2 version. This creates
symlinks to previously defined Docker image versions, uses the
same version of container images used for k8s 1.29, k8s 1.30
and k8s 1.25 and uses the same volume-snapshot-controller
of K8s 1.26 and CRDs of k8s 1.25.

Additionally, this change updates the symlinks for 1.31 to point
to the images used in 1.30.

Test Plan:
PASS: Install ISO with k8s 1.32.2 on AIO-SX
PASS: Install ISO with k8s 1.32.2 on AIO-DX
PASS: Install ISO with k8s 1.32.2 on Standard

Story: 2011340
Task: 51980

Depends-On: https://review.opendev.org/c/starlingx/integ/+/947189

Change-Id: I1bea4226f53024b74d60a98ffc6afabcbea4d412
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>