Avoiding subcloud reboot when enrollment seed iso is detected
Currently, the enrollment process requires a reboot of
the subcloud to start cloud-init and reconfigure the
OAM interface.

This commit adds changes to pass an excluded_operations
list to rvmc_install.py script with which predefined
operations can be omitted, in this case the reboot
during the enroll-init stage. This way, passing the
'poweroff_host' operation prevents the subcloud from
being turned off and back on to initiate enrollment.

After this change the long wait for the host reboot can
be avoided, the wait time for sysinv endpoint open can
be reduced into 4 minutes.

Test Plan:

PASS: - Complete the factory install on the subcloud.
      - Run the enrollment process through dcmanager and
        verify that it completes successfully.
      - Verify that the subcloud is not rebooted at the
        enroll-init stage. The enrollment should complete
        without the need to reboot the subcloud.

Depends-On: https://review.opendev.org/c/starlingx/utilities/+/947441

Closes-bug: 2109795

Change-Id: Iab7e353201d9ce377dab5b9c3c30cbf6a85e4343
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>
Co-authored-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>