Updated restore sw-deployments to ignore unwanted items
During a restore of a System Controller that had a N-1 deployment
loaded, restore was failing because of how the metadata was arranged.

This fix, allows restore to correctly detect the metadata and also
exclude it from being restored.  When the N-1 load is version 22.12 the
deployment cannot be restored because it is in a legacy format.  To get
the N-1 load back, the load must be reimported after restore is
complete.

TEST PLAN
PASS: B&R System controller with N-1 loaded
PASS: B&R System controller with N-1 loaded and deployed patches
PASS: B&R System controller with only deployed patches
PASS: B&R System controller with nothing

Closes-Bug: 2110224
Change-Id: I43cb31ea8d16f8568ed2db1891c34eb11211cc65
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>