Improve the prestage to configure the release state appropriately
The current implementation compares commits between sysroot
and feed repo to determine the "deployed" status of releases
as long as the commit history is the same between both.

It may happen that the commit history differs between sysroot
and ostree, causing the release status to not be set
appropriately.

This commit improves the prestage implementation to set the
release state based on the current state in the subcloud,
in case the history changes.

Test Plan:
PASS: - Bring up a SystemController and subcloud with
        24.09
      - Deploy patches 24.09.1, 24.09.2 on System
        Controller
      - Prestage subcloud for-sw-deploy with 24.09
      - Deploy prestaged patches 24.09.1, 24.09.2 on
        Subcloud
      - Rollback System Controller to 24.09
      - Deploy patches 24.09.1, 24.09.2 and 24.09.3 on
        System Controller.
         At this point, the ostree history on the
         System Controller changes, so in the next step
         it will be synchronized with the Subcloud,
         leaving the Subcloud with a different commit
         history than it had.
      - Prestage subcloud with 24.09
      - Verify the state of the patches on subcloud.
        Patches 24.09.1, 24.09.2 should remain in
        "deployed" state.
        Patch 24.09.3 should have the "Available" state.
      - Apply patch 24.09.3 to the subcloud and verify
        that it is synchronized with the System
        Controller.

Closes-bug: 2110570

Change-Id: I7ef325e513f70a5c71cd1b9bf9cd9e142bde4b02
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>