Fix rehome using the mgmt network
After the admin address reduction feature was implemented, rehome
without the admin network (mgmt only) started to fail.

This change fixes wrong variables usage in this case.

Test plan:
- rehome using mgmt network

Note: It was not possible to finish all rehome tests because there
was an unrelated cert error. This fix needs to go in first so that any
other errors can be fixed later.

Story: 2011191
Task: 52222

Change-Id: I6f069f991e4bbf6760382515c1072a77ad759528
Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>