Fix wipe_osds.sh script to skip read-only disks
During a subcloud bootstrap, the wipe_ceph.sh script failed when
checking a USB device identified as TYPE='disk' but set as
read-only.

This script locks the device to avoid race conditions with udev
and updates to partition data/metadata. And it fails when trying
to lock a read-only device.

The script is changed to check if the device is set as read-only.
If it is set to read-only, it will be skipped to avoid this error.
Ceph disks should never be detected as read-only devices.

Example of 'lsblk' command output to detect the TYPE and RO values:
lsblk --nodeps --pairs /dev/sdb
    NAME="sdb" MAJ:MIN="8:16" RM="1" SIZE="472K" RO="1" TYPE="disk" \
    MOUNTPOINT=""

Test-Plan:
  PASS: AIO-SX: install the system, and before bootstrapping, set a
        disk to be read-only. Run the boostrap and check if that disk
        was skipped when running the wipe_osds.sh script.

Closes-bug: 2111580

Change-Id: If657356fe3f1c8480d4d26122312d10ea3486261
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>