Set the appropriate metadata file during the prestage for-sw-deploy
The current code, when performing a for-sw-deploy
prestage, synchronizes the ostree repository and
the System Controller metadata. However, if a
patch is already deployed to the subcloud,
instead of synchronizing the System Controller's
metadata, it uses the existing one in the subcloud.
This causes the deployment to fail at some stages.

This commit adds the change to force the use of
the existing metadata in the System Controller,
thus synchronizing the subcloud.

Test Plan:
PASS - Deploy a SystemController with 24.09
     - Deploy a second SystemController with a
       subcloud with different 24.09 build date.
     - Deploy a patch on first SystemController
     - Rehome the subcloud to first SystemController
     - Prestage the subcloud for-sw-deploy
     - Verify that the patch is prestaged on the
       subcloud
     - Run software precheck to verify the prestaged
       patch

Closes-bug: 2110570

Change-Id: I312f9ca3be40cc5f38b74f509304673f2a8b704a
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>