Increase timeout for restarting networking during restore
The speed of the networking restart is determined by the type and number
of interfaces on the system.  Especially bonds, which have a 10 second
sleep for stabilization.  For example, on a system with 6 bonds, the
minimum time to restart networking will be over 60 seconds.

This change bumps the timeout to 120 seconds.  A more deterministic
improvement should follow.

TEST PLAN
PASS: AIO-SX optimized restore
  * Unaffected system, no impact
PASS: AIO-DX legacy restore
  * Affected system, issue resolved

Partial-Bug: 2112117
Change-Id: I6b60fc2fdc0311e3c6630ead1848b75e65277eb0
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>