Remove patching user from rehome
After [1], the patching user was removed from users_to_replicate.
Because of this, it's necessary to remove the calls that access its
value during rehoming.

Test plan:
  - PASS: Rehome a subcloud and verify the playbook completed
          successfully.

Note: There's currently a known issue with rehoming, in the task
"Check admin-ep-cert.pem updated". For this test, the task was
commented out.

[1]: https://review.opendev.org/c/starlingx/distcloud/+/949363

Closes-bug: 2114907

Change-Id: Iae2c5cfb7488fbaf3f91d395312fc14a32028e70
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>