Add new MariaDB 'sys' DB to ignore list
After the upversion to Caracal, the STX-Openstack MariaDB
new version creates a database named 'sys' during the MariaDB
pod creation. Because of this DB the restore process fails
with an error saying that the sys database already exists
when ansible tries to create the DB in the MariaDB cluster.

This review add the 'sys' database to the list of DBs to be
ignored during the backup creation. This fix will remove the
'sys' DB dump from the backup tarball and fix the error in the
restore

Test Plan:
PASS - Run ansible backup script
PASS - Verify that the sys database dump was not created
PASS - Restore STX-Openstack with the backup file generated
PASS - Verify all Openstack resources were restored
PASS - Verify that STX-O functions as it should after restore

Closes-Bug: #2115271

Change-Id: I36953ce2cdfdf664cb3bf4d808dfd4f64275c6ee
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>