Add factory-install files/folders to BnR
This commit adds the /var/lib/factory-install directory and
/etc/udev/rules.d/99-seediso.rules file to the platform backup and
restore. The files/folders will only be backed up in a factory
installed system prior to enrollment, as they're only relevant during
this period.

This is necessary for subcloud enrollment post restore of factory
install, as enroll need to know the completed stages and scripts
to correctly execute.

Test plan:
  - PASS: Run the backup in a factory installed system and verify
          the necessary files are present in the generated tarball.
  - PASS: Run the factory restore through dcmanager and verify the
          necessary files present and with the same content as before.
  - PASS: Finish subcloud enrollment, take a new backup and then
          restore it. Verify the /var/lib/factory-install folder and
          /etc/udev/rules.d/99-seediso.rules wasn't restored.

Story: 2011454
Task: 52418

Change-Id: Ib98f994232a3061c1864d531e1f80ec440823cfd
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>