Restore ldap user accounts in optimized restore
In optimized restore, in the task "Restore ldap" the command
"sudo slapadd -F /etc/ldap/schema -l /opt/backups/ldap.db" is
failing because it is trying to add data to the existing db
which is conflicting with existing key/data pair and throwing
an error "key/data pair already exists",
For the update to work, the contents of the directory
/var/lib/ldap should be cleared, this task is under the role
"restore-platform" which is not called by optimized restore.
This change moves the following tasks
   - Delete ldap directory
   - Recreate ldap directory
to backup-restore role thus clearing the db contents before
updation.

Test Cases:
PASS: Perform an optimized restore and verify the ldap user
      accounts are successfully restored
PASS: Perform the legacy restore and verify that ldap user
      accounts are successfully restored.
PASS: Verify that system is successfully bootstrapped with
      the changes

Closes-Bug: 2116966

Change-Id: I43bb973eded9550bdf72bfc35c5275d48395439b
Signed-off-by: amantri <ayyappa.mantri@windriver.com>