Rework patch during enrollment for chain dependencies
The patch-before-enroll only checks the highest patch id from the
system controller and get its dependency to upload and apply in the
subcloud. This dependency check is insufficient to get all the
dependencies in chain.

This commit fixes this issue by checking the dependencies in chain,
ensures all the dependencies can be applied during enrollment. As the
check of the software on the system controller local should not be hit
too often that may cause the API service being overwhelmed if
enrolling a large amount of subclouds in parallel, this commit switch
to software client to get all the releases' info in one request.

This commit also fixes an issue that "state" keyword can be existed in
the patch description using regex.

Test plan:
1. Passed - verify subcloud enrollment successful with patch 24.09.210
   on subcloud running 24.09.200.
2. Passed - verify check patch failed with patch 24.09.300 but missing
   patch 24.09.210 against subcloud running 24.09.200
3. Passed - verify check patch failed if subcloud running 24.09.100,
   having .210, .300 but does not have .100 in the system controller.
4. Passed - verify patch skipped if no patch higher than subcloud's
   24.09.100 version.
5. Passed - upload .200, .210, .300 patches to the system controller,
   verify all the three patches are uploaded to the 24.09.100 for
   patch.

Closes-bug: 2120992

Signed-off-by: Jiang, Yuxing <Yuxing.Jiang@windriver.com>
Change-Id: I028313b5ceaff03b27d112bab1dd3d79b9cff98d