Remove snapshots during BnR with wipe_ceph_osds=true
During backup and restore, when there are snapshots provisioned by
ceph, they are not deleted, although the OSDs are wiped and the app
is removed/applied, which will cause issues when using them or if
they are already being used by a pod.

To fix this, when a PVC is identified as having been created from
a Ceph snapshot, both the PVC and the snapshot itself are removed,
since all data has been wiped. Therefore, if any pod is using the
PVC, that PVC must be provisioned, as the pod will be in the
"ContainerCreating" status, waiting for the PVC.

Test Plan:
 - PASS: B&R on AIO-SX, with rook-ceph applied, with rbd and
         cephfs snapshots created, and wipe_ceph_osds=true.
         Checks whether PVCs created from snapshots and the
         snapshots themselves have been deleted.

 - PASS: B&R on AIO-SX, with platform-integ-apps applied, with rbd
         and cephfs snapshots created, and wipe_ceph_osds=true.
         Checks whether PVCs created from snapshots and the
         snapshots themselves have been deleted.

Closes-Bug: 2121518

Change-Id: Ib2934d35342d12035abd18fe0fa4ca2eae605748
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>