Fix rook-ceph BnR with snapshot
During rook-ceph recovery, there's a cephfs recovery step. However,
when a cephfs snapshot is created, the reference to that snapshot
is lost. Therefore, if a pod uses a PVC created from a snapshot or
uses that PVC after the restore, MDS will report a metadata error
because it can't find that reference.

To resolve this, during cephfs recovery, the existence of a
snapshot is checked and then the 'cephfs-data-scan scan_links'
command is run to scan the links between different objects
in the CephFS filesystem and update the metadata database with
this information.

Example command output:
"""
mds.0.snap  updating last_snap 1 -> 2
mds.0.snap  updating snaps {{}} -> {{2=snap(2 0x10000000002 \
'csi-snap-a08f11c0-65e2-42bd-bbe4-e72862e9f5ad' 2025-08-22T16:23:27.312001+0000)}}
"""

Test Plan:
 - PASS: B&R on AIO-SX/AIO-DX, with rbd and cephfs snapshots created.
 - PASS: Checks that all pods are 'running' and PVCs are 'bound'.

Closes-Bug: 2121525

Change-Id: I2d147198bc2bf91a0a185e837df898ba03bcff50
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>