Update monitor-enroll-init-ipmi failure log
This commit updates the monitor-enroll-init-ipmi role so that if in a
future this role is imported by other playbooks, the
operation that imports it can provide its own string through the
`operation_string` variable. This way, in a failure case, the
operation will be clearly distinguished in the error message.

Test Plan:
PASS: Run the enroll_init.yml and induce a failure. Verify that the
      enroll-init operation is printed.
PASS: Run a new playbook with a different operation_string that
      imports this role, and induce a failure. Verify that the custom
      string is printed.

Story: 2011455
Task: 52605

Signed-off-by: Enzo Candotti <Enzo.Candotti@windriver.com>
Change-Id: I417ee8fb5593f2206137d050fba2e7d497b3c7a9