Force the re-transfer of feed repo for N-1 subcloud prestage
Currently, when prestaging is performed for an N-1 subcloud,
the entire feed repo is transferred to the subcloud.
However, if the prestage process is interrupted (e.g., due to
a network issue or subcloud reboot), the transfer may be
incomplete. In such cases, if the prestage is retried, the
existing (but incomplete) feed repo is detected, and the
transfer is skipped. This results in the release not being
fully available on the subcloud.

To prevent this scenario, instead of checking if the
directory exists in N-1 subcloud, the entire feed repo is
deleted and forced to be re-transferred. This way, if the
transfer completes successfully, the release can be deployed.
If for some reason the transfer fails, the feed repo will be
recreated on the next retry.

Additionally, a change has been introduced to correctly
display prestaged releases in for-install. The current code
searches for all releases in platform-backups, and in the
case of an N-1 subcloud that has never been prestaged in
for-install, it will also display its release, which is
incorrect because only prestaged releases should be displayed.
To prevent this, for each prestage run in for-install, a flag
file is created to identify the release in question.
If that file is found in the release's directory (created when
for-install is run), it indicates that it should appear in the
list of prestaged releases. Otherwise, if the file doesn't
exist, it is assumed that the release is a pre-existing
version and should not be displayed in the list.

Test Plan:
PASS - Prestage a new N-1 subcloud for-sw-deploy
     - Verify that the subcloud is correctly identified as
       N-1
     - Verify that the feed repo is fully transferred
     - Verify that the release appears in the list of
       prestaged releases for-sw-deploy
     - Verify that the N release is in available state

PASS - Re-run the prestage for-sw-deploy on the same N-1
       subcloud
     - Verify that the feed repo is deleted and fully
       re-transferred
     - Verify that the release appears in the list of
       prestaged releases for-sw-deploy
     - Verify that the N release is in available state

PASS - Prestage a new N-1 subcloud for-sw-deploy
     - In the middle of the transfer, simulate a network
       interruption, or reboot the subcloud
     - Once the subcloud is back online, re-run the
       prestage for-sw-deploy.
     - Verify that the feed repo is fully transferred
     - Verify that the release appears in the list of
       prestaged releases for-sw-deploy
     - Verify that the N release is in available state

PASS - Upload and apply a patch on SystemController
     - Prestage an existing N subcloud for-sw-deploy
     - Verify that the new patch is in available state

PASS - Prestage a new N-1 subcloud for-install
     - Verify that the N release appears in the list of
       prestaged releases for-install
     - Verify that the N-1 release does not appear in
       the list of prestaged releases for-install

Closes-bug: 2123869

Change-Id: I3ca00cfd68e08278024b77095ae0cf865ec31084
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>