Ensure backup uses active kube version after upgrade
When the user upgraded the equipment to a new StarlingX version and
initiated a system backup, the code verified whether the current
Kubernetes version was the latest supported one by the new
StarlingX release. If the Kubernetes version was outdated,
the user was prompted to upgrade it.

The verification that was intended to ensure that backups were
compatible with the new environment and to prevent errors during the
restore process due to an outdated Kubernetes version caused an
issue when the system was upgraded to the latest version. It
recommended using the latest version but did not pass the active
Kubernetes version to the backup process, instead using the
overridden version from installation.

This change updated the logic so that backup and restore operations
worked correctly with the latest Kubernetes versions.

Test Plan:
PASS: On a AIO-SX, Perform k8s upgrade to latest version,
      perform BnR , verify that active version is passed and
      BnR is successfull.
PASS: Install an iso with k8s 1.29 , perform backup and verify that
      backup fails for older k8s version.
PASS: Install AIO-DX with k8s 1.29, perform k8s upgrade to 1.32 then
      perform legacy backup and restore.
PASS: AIO-SX, Perform ansible replay.

Closes-Bug: 2122272

Change-Id: Ib443674d855d7afeb90bc4318f2b55a2bbfecdb0
Signed-off-by: Saba Touheed Mujawar <sabatouheed.mujawar@windriver.com>