Add support for auto-restore on pre-installed subclouds
This commit allows triggering an auto-restore by mounting a cloud-init
seed iso containing the backup files. The playbook is modified to use
the RVMC script to mount the seed iso if a rvmc_config_file is defined.

It allows running the auto-restore operation without --with-install on
an already installed (but not bootstrapped) subcloud.

Test Plan:
01. PASS - Install a subcloud with a prestaged ISO and then create a
    central subcloud backup. Manually reinstall the prestaged ISO and
    then run the auto-restore without both --with-install and
    --local-only. Verify that the subcloud is restored using the
    bundled central backup file without automatic reinstallation.

Story: 2011454
Task: 52796

Change-Id: I57ed513ee3adf72d7e63608e94ecfd9d27f53d75
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>