Truncate helm overrides files to ensure auto-reapply
During BnR with the rook-ceph backend configured, the application
must be reapplied after the 'system restore-complete' command to
ensure the correct app configuration. However, for this to happen,
the md5sum hashes of the overrides must be different, which is done
through 'evaluate_app_reapply' in the conductor.

To fix this, after rook-ceph recover, the contents of the helm chart
overrides are erased, leaving only the empty *.yaml files.

Test Plan:
 - PASS: B&R on AIO-SX with rook-ceph backend configured
 - PASS: B&R on AIO-DX with rook-ceph backend configured

Partial-Bug: 2122729

Change-Id: I2279b21eaf14f566268cd042f1563a70d134be8c
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>