Ignore errors to allow playbook continuation
This commit ignores errors if any, during extraction of
local_registry_filesystem.tgz file. This will allow the bootstrap
playbook to continue even if the file is corrupted, as docker images
can be pulled from configured sources.

For partial extraction of the tar file, when pulling an image from the
local registry fails, it was giving an error message and exiting.
Code has been modified to fallback to pulling from remote registry
in such case.

Test Plan:
PASS: Run subcloud redeploy and verify the bootstrap playbook is not
      stopped, with help of corrupted empty
      local_registry_filesystem.tgz file
PASS: Verify bootstrap playbook execution not stopped for corrupted
      local_registry_filesystem.tgz file with partially extracted
      registries

Closes-Bug: 2125111
Change-Id: I8fb14cc25dc4c22cb82ab65007335d682de719a3
Signed-off-by: sshathee <shunmugam.shatheesh@windriver.com>