Sync updates packages for prestage for-sw-deploy
Current prestage for-sw-deploy workflow does not sync the updates
packages directory to the subcloud when prestaging an N-1 release, which
is required by USM to perform the deploy, causing the deploy to fail.

This commit adds a new task file to sync the updates packages directory
to the subcloud when prestaging an N-1 release. To do so, the
restrictions to execute the sync-software-metadata role are modified
to allow the execution of the new task file when prestaging an N-1
release. For N release prestage, the existing restrictions still apply.
Additionally, if the presage software version is lower than the current
subcloud software version, the prestage is aborted because we dont want
to prestage an N-1 release to an N subcloud.

Test Plan:
PASS - Prestage N-1 release which is in unavailable state.
     - Verify that the updates packages directory is synced to the
       subcloud.
     - Verify that the prestage of the N-1 release completes
       successfully and the release state is Available
     - Verify that the deploy of the N-1 release completes successfully.

PASS - Prestage N release which is in deployed state.
     - Verify that the prestage of the N release completes successfully
       and the release state is Available
     - Verify that the deploy of the N release completes successfully.

PASS - Prestage N-1 release to an N subcloud
     - Verify that the prestage is aborted with the appropriate error
       message.

Regression Test:

    - Prestage N release on N-1 subcloud for deployment
    - Prestage N release on N subcloud for deployment
    - Prestage N release on N-1 subcloud for install
    - Prestage N-1 release on N-1 subcloud for install
    - Prestage N release on N subcloud for install
    - Prestage N-1 release on N subcloud for install

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/961844

Closes-Bug: 2125230

Change-Id: I6b840877180bd4c301627694cc98670001567373
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>