Fix rook-ceph restore
During the rook-ceph BnR improvements [1], a variable was lowercase
instead of uppercase, which caused the /etc/ceph/ceph.conf created
on the recovery pod to be without the monitor IP, only with the
port (':6789'), causing a lack of communication with ceph. This
issue only happens when a pod on another host (not the target host)
has ceph provisioned when the monstore rebuild pod is running.

Additionally, it was also observed that during the patching of the
"rook-ceph-recovery" configmap that contained the error due to lack
of communication, it contained characters that crashed the command
in kubectl. Therefore, the patch was implemented using "jq" to fix
these "escaping" issues.

[1]: https://review.opendev.org/c/starlingx/ansible-playbooks/+/953340

Test Plan:
 - PASS: On Standard, with dedicated deployment model,
         with monitor on controller-1, compute-0 and compute-2.

Closes-Bug: 2126668

Change-Id: If4753799a4c268549b0edb48c5afdf3cae6c4add
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>