Update secrets with registry credentials
As part of rehoming, the keystone password for sysinv is updated for
the new SystemController value. A callback then comes from keystone
to sysinv in which the secrets that contain the password used to
authenticate to the docker registry for each namespace are updated.

However, since the playbook needs to restart sysinv right away,
there is a racing condition between this callback and the sysinv
restart. To prevent this, this change adds a script that makes a
RPC call to sysinv, to run the audit that updates the credentials
in the secrets in a synchronous way (w/o waiting for the callback).

PASS: Rehome a SX subcloud
      Confirm that the registry secrets were updated

PASS: Enroll SX subcloud

Depends-on: https://review.opendev.org/c/starlingx/config/+/963219

Closes-bug: 2126680

Change-Id: If595916b87e2c0cbfce60721547ea74d6ac4b168
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>