Allow stx-openstack restore from different versions
The openstack app restore is limited to only allow restoring
a backup file made on the same version as the current host.
This limitation is due to the use of the 'software_version'
variable that gets the version of the running host.

This change creates a variable named 'backup_software_version'
that reads the software version from the backup tarball to
avoid the issue.

The 'warn: false' was added to avoid a warning suggesting the
use of unarchive instead of tar. Unarchive cannot be used on
this context, as it doesn't provide a read-only option to
list the files

This change does NOT affect platform backup and restore as this
playbook is used exclusively on openstack app restore.

Test plan:

PASS: build-pkgs -c -p playbookconfig
PASS: build-image
PASS: backup and restore openstack app with the same software
versions
PASS: backup and restore openstack app with different software
versions

Closes-Bug: 2127692

Change-Id: Ifd622723947bbc47972244ce8ab737be09eb1f88
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>